update nav menu with about button
